\name{express}
\alias{express}
\docType{data}
\title{Simulated dataset with 2000 features}
\description{
Simulated dataset with 2000 features containing a four-way interaction term. The simulated outcome y = x1578*x1430*x692*x1223 }
  \usage{data(express)}
\details{Simulated.
}
\examples{
  data(express)
}
\source{
Ishwaran H., Kogalur U.B., Chen X. and Minn A.J. (2011). Random survival forests for high-dimensional data. \emph{Statist. Anal. Data Mining}, 4:115-132.
}


