\name{HIvimp}
\alias{HIvimp}
\title{High Order Interaction Detection and Variable Importance}
\author{Yifan Sha and Min Lu}
\usage{HIvimp(formula, data, method = c("hivimp", "acuvimp")[1], terms = NULL,
       PMDorder = TRUE, importance = "permute", block = 1,
       subRF = TRUE, subvars = function(data){ceiling((ncol(data)-1)/5)},
       w.initial = "vimp",
       wt = function(btpmd,digpmd){log(1/btpmd/digpmd)},
       itrSub = 5, wtSub = TRUE,
       itrWt = 1,
       subQtl = function(btpmd,digpmd){
                which(btpmd < quantile(btpmd, probs = 0.1))},
       nHIvimp = function(all.cmbn){min(length(all.cmbn),372)},
       verbose = TRUE)}
\description{This function detects and provides ranking criteria for high order interaction terms.}
\arguments{
\item{formula}{A symbolic description of the model to be fit.}
\item{data}{Data frame containing the y-outcome and x-variables.}
\item{method}{Method(s) used for ranking interaction terms. Choose ``hivimp" for high order interaction variable importance or/and choose ``acuvimp" for exact high-order interaction deviance.}
\item{terms}{The highest order interaction terms up to search. The default value NULL indicates automatic detection using function ``nterm"."}
\item{PMDorder}{Whether function ``pmdhvp" is used for fast ranking the first ``nHIvimp" interaction terms before passing to ``hvp" function to decide the final ranking.}
\item{importance}{Method for computing variable importance (VIMP). It is the same as ``importance" in \emph{rfsrc} function in \emph{randomForestSRC} package. }
\item{block}{Specifies number of trees in a block when calculating VIMP. It is the same as ``block.size" in \emph{vimp} function in \emph{randomForestSRC} package.}
\item{subRF}{logical. Should variable selection be conducted using subRF method (see ``var.sel" function)? Strongly recommended if p > 50}
\item{subvars}{Number of variables selected for fitting each random forest. Default method uses all variables if p < 50, and ceiling(p/ln(p)) otherwise.}
\item{w.initial}{A probability vector, according to which features are selected (see w0 in subRF function) for each subRF. When equals to ``vimp", then variable importance is used. When equals to ``md", then minimal depth from maximal subtree is used.}
\item{wt}{A function calculating variable weights using PMD matrix where \emph{btpmd}[i] is the
average of \emph{i}th row in the PMD matrix where only off diagonal elements that have smaller values are used and \emph{digpmd}[i] is the \emph{i}th diagonal element in the PMD matrix.}
\item{itrSub}{Number of subRFs to grow. Default is set as terms*log10(p)}
\item{wtSub}{logical. Should weighted random forests grown?}
\item{itrWt}{Number of iterations of each weighted subRF.}
\item{subQtl}{Which features will be selected from each subRF. ``digpmd” is the diagonal elements of the PMD matrix and
``btpmd” is the average of \emph{i}th row in the PMD matrix used
where only off diagonal elements that have smaller values are used.}
\item{max.var}{Cut to threshold variables as final variable selection based on frequency table generated by variables selected from all subRFs.}
\item{nHIvimp}{Number of iteractions that will be firstly selected using ``pmdhvp". It is a function of a vector ``all.cmbn", where each PMD high order interaction importance is stored.}
\item{qtlscr}{Perform normalized high order interaction variable importance density estimation for quantile score computation.}
\item{verbose}{Set to TRUE for verbose output.}
}

\value{
\item{terms}{The highest order interaction terms up to search.}
\item{interaction}{A dataframe with ranking criteria for each interaction term in the row and several methods in the column. ``joinvimp" for joint vimp; ``HIvimp" for high order interaction variable importance; ``nmHIvimp" for normalized high order interaction variable importance; ``HIvimp2" for high order interaction variable importance where individual vimp uses ``pmdhvp"; ``nmHIvimp2" for normalized ``HIvimp2"; ``acuvimp" for exact high-order interaction deviance; ``joinvimp.sub" for reinforced joint vimp and so on.}
\item{extracted.nm}{A list of seperated variable names in the interaction terms in ``interaction".}
\item{rf.obj}{Random Forest object fitted with selected variables.}
\item{pmd}{PMD matrix from rf.obj.}
\item{var.sl}{Names of variables selected.}
\item{var.sl.list}{A list of each subRF's output: the \emph{j}th object is the output of the \emph{j}th subRF with two elements: \emph{sl.w} contains the weights (calculated using input \emph{wt}) of selected variable and \emph{sl.var} contains the names of selected variable. Final result (\emph{var.sl}) is the union of each set \emph{sl.var}.}
\item{itrwt}{A list of each weighted subRF's weights: the \emph{j}th object is the weight path of the \emph{j}th subRF with \emph{itrWt} elements. See the output of \emph{wt.itr} function. }
\item{pmdvimp.long}{ PMD variable importance for each interaction terms in \emph{inter}.}
}


\references{
Ishwaran H. (2007). Variable importance in binary regression trees and forests, \emph{Electronic J. Statist.}, 1:519-537.

Ishwaran H., Kogalur U.B., Gorodeski E.Z, Minn A.J. and Lauer M.S. (2010). High-dimensional variable selection for survival data. \emph{J. Amer. Statist. Assoc.}, 105:205-217.

Ishwaran H., Kogalur U.B., Chen X. and Minn A.J. (2011). Random survival forests for high-dimensional data. \emph{Statist. Anal. Data Mining}, 4:115-132.
}

\examples{
data(express)
o <- HIvimp(y~., data = express[,1:100], method = "hivimp", terms = NULL, PMDorder = TRUE,
             importance = "permute", block = 1,
             subRF = TRUE, subvars = NULL,
             w.initial = "vimp",
             wt = function(btpmd,digpmd){log(1/btpmd/digpmd)},
             itrSub = NULL, wtSub = TRUE,
             itrWt = 1, subQtl = NULL, max.var = 50,
             nHIvimp = function(all.cmbn){min(length(all.cmbn),50)},
             qtlscr = TRUE,
             verbose = TRUE)
head(o$interaction)
}
